Exemplos add.

+++ runCheck.sh
@@ -22,6 +22,10 @@ INPUTDIF3DEX2='ex2_3D_0_ls.dat    ex2_3D_1_ls.dat
                ex2_3D_0_ggln.dat  ex2_3D_1_ggln.dat  
                ex2_3D_0_gglc.dat  ex2_3D_1_gglc.dat' 
 
+INPUTDIF3DEX3='  ex3_3d_hexa_0_ls.dat   ex3_3d_hexa_1_ls.dat ex3_3d_hexa_2_ls.dat  
+               ex3_3d_hexa_0_ggln.dat ex3_3d_hexa_1_ggln.dat ex3_3d_hexa_2_ggln.dat 
+               ex3_3d_hexa_0_gglc.dat ex3_3d_hexa_1_gglc.dat ex3_3d_hexa_2_gglc.dat'  
+
 
 NAMEBIN=mfvCell_gnu_O3
 
@@ -80,7 +84,17 @@ if [ $? == 1 ];then
   exit 1
 fi
 
+# difusao 3D
+tar -xvzf test/input/dif/non_orthogonal/3D/*.tar.gz -C test/binTest
+cp test/input/dif/non_orthogonal/3D/GreenGaussCell/*.dat test/binTest
+cp test/input/dif/non_orthogonal/3D/GreenGaussNode/*.dat test/binTest
+cp test/input/dif/non_orthogonal/3D/LeastSquare/*.dat test/binTest
+test/binTest/run_dif_exato_3D_Ex3.sh "$INPUTDIF3DEX3" "$NAMEBIN" 
+if [ $? == 1 ];then
+  exit 1
+fi
+
 DIR="test/binTest"
-rm  $DIR/*.dat $DIR/*.vtk  $DIR/*.txt   $DIR/*.csv $DIR/$NAMEBIN $DIR/*.mtx
+rm  $DIR/*.dat $DIR/*.vtk  $DIR/*.txt   $DIR/*.csv $DIR/$NAMEBIN $DIR/*.mtx 
 
 exit 0
